requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Pillow>=10.0.0
